Add files via upload
This Crestron program will use a Simpl+ module and SIMPL# Library to load my Flickr feed pictures and display them on a Xpanel one at a time. I have modified the SIMPL# Library in C# to change the timing from 60 seconds to 15 second per frame.
Note: There are only 3 pictures in my Flickr feed test account, so once the 3rd is displayed the show ends and stays on that frame.
This is a example from the Introduction into SIMPL# from Crestron. You require Visio Studio 2008 SP1 to make and commit edits. Updating the VS 2008 to any newer version will not work.